server/requirements.txt
Flask==2.2.3
psycopg2-binary==2.9.6
matplotlib==3.6.3
